requirements.txt
pdfminer.six==20200517
pdfplumber==0.5.24
PyMuPDF==1.18.4
